+++ includes/classes.php
@@ -1,20 +1,20 @@
-<?php defined( 'PROPRO_VERSION' ) || die;
+<?php defined( 'PRDWC_VERSION' ) || die;
 
 require(plugin_dir_path(__FILE__) . 'post-type-project.php');
 
-function propro_is_project_product($product_id) {
+function prdwc_is_donation($product_id) {
   // return true; // let's handle this later
 	return (wc_get_product( $product_id )->get_meta( '_linkproject' ) == 'yes');
 }
 
-function propro_allow_custom_amount() {
-	if(get_option('propro_custom_amount') == 'yes') return true;
+function prdwc_allow_custom_amount() {
+	if(get_option('prdwc_custom_amount') == 'yes') return true;
 	return false;
 }
 /**
- * [PROPRO description]
+ * [PRDWC description]
  */
-class PROPRO {
+class PRDWC {
 
   /*
   * Bootstraps the class and hooks required actions & filters.
@@ -47,32 +47,32 @@ class PROPRO {
     add_filter( 'woocommerce_product_add_to_cart_text', __CLASS__ . '::add_to_card_button', 10, 2);
     add_filter( 'woocommerce_product_single_add_to_cart_text', __CLASS__ . '::single_add_to_card_button', 10, 2);
 
-		if(get_option('propro_create_project_post_type', false) ==  'yes')  add_action( 'init', 'propro_register_project_posttype' );
+		if(get_option('prdwc_create_project_post_type', false) ==  'yes')  add_action( 'init', 'prdwc_register_project_posttype' );
 
 		add_action( 'plugins_loaded', __CLASS__ . '::load_plugin_textdomain' );
   }
 
   static function load_plugin_textdomain() {
 		load_plugin_textdomain(
-			'project-products',
+			'project-donations-wc',
 			false,
 			dirname( dirname( plugin_basename( __FILE__ ) ) ) . '/languages/'
 		);
 	}
 
   static function add_to_card_button( $text, $product ) {
-    if($product->get_meta( '_linkproject' ) == 'yes') $text = __('Support project', 'project-products');
+    if($product->get_meta( '_linkproject' ) == 'yes') $text = __('Donate', 'project-donations-wc');
   	return $text;
   }
 
   static function single_add_to_card_button( $text, $product ) {
-    if($product->get_meta( '_linkproject' ) == 'yes') $text = __('Support project', 'project-products');
+    if($product->get_meta( '_linkproject' ) == 'yes') $text = __('Donate', 'project-donations-wc');
   	return $text;
   }
 
   static function add_to_cart_message( $message, $product_id ) {
       // make filter magic happen here...
-      if(!empty($_POST['propro-project-name'])) $message = sanitize_text_field($_POST['propro-project-name']) . ": $message";
+      if(!empty($_POST['prdwc-project-name'])) $message = sanitize_text_field($_POST['prdwc-project-name']) . ": $message";
       return $message;
   }
 
@@ -80,8 +80,8 @@ class PROPRO {
     $product_type_options['linkproject'] = array(
       "id"            => "_linkproject",
       "wrapper_class" => "show_if_simple show_if_variable",
-      "label"         => __('Project Product', 'project-products'),
-      "description"   => __('Check to add a project field to product page.', 'project-products'),
+      "label"         => __('Project Donation', 'project-donations-wc'),
+      "description"   => __('Check to add a project field to product page.', 'project-donations-wc'),
       "default"       => "no",
     );
     return $product_type_options;
@@ -94,9 +94,9 @@ class PROPRO {
 	static function display_custom_fields_simple() {
 		global $post, $product;
 		if( $product->has_child() ) return; // already called by woocommerce_before_variations_form
-		if(!propro_is_project_product( $post->ID )) return;
+		if(!prdwc_is_donation( $post->ID )) return;
 
-		PROPRO::display_custom_fields();
+		PRDWC::display_custom_fields();
 	}
 
 	static function display_custom_field($args) {
@@ -158,7 +158,7 @@ class PROPRO {
 
 			default:
 			$input = sprintf(
-				'<input type="%1$s" class="input-text" name="propro-%2$s" value="%3$s" placeholder="%4$s" %5$s>',
+				'<input type="%1$s" class="input-text" name="prdwc-%2$s" value="%3$s" placeholder="%4$s" %5$s>',
 				$type,
 				$name,
 				$value,
@@ -167,9 +167,9 @@ class PROPRO {
 			);
 		}
 		printf(
-			'<div class="propro-field propro-field-%1$s">
+			'<div class="prdwc-field prdwc-field-%1$s">
 				<p class="form-row form-row-wide">
-					<label for="propro-%1$s" class="%2$s">%3$s%4$s</label>
+					<label for="prdwc-%1$s" class="%2$s">%3$s%4$s</label>
 					%5$s
 				</p>
 			</div>',
@@ -187,35 +187,35 @@ class PROPRO {
   */
   static function display_custom_fields() {
 		global $post, $product;
-		if(!propro_is_project_product( $post->ID )) return;
+		if(!prdwc_is_donation( $post->ID )) return;
 
 		$project = (isset($_REQUEST['project'])) ? sanitize_text_field($_REQUEST['project']) : NULL;
 		$project_id = (isset($_REQUEST['project-id'])) ? sanitize_text_field($_REQUEST['project-id']) : NULL;
 		$price = $product->get_price();
 		$amount = sanitize_text_field((isset($_REQUEST['amount'])) ? $_REQUEST['amount'] : ((isset($_REQUEST['nyp'])) ? $_REQUEST['nyp'] : NULL));
 
-		$project_post_type = (get_option('propro_create_project_post_type') == 'yes') ? 'project' : get_option('propro_project_post_type');
-		printf('<div class="propro-fields" style="margin-bottom:1rem">');
+		$project_post_type = (get_option('prdwc_create_project_post_type') == 'yes') ? 'project' : get_option('prdwc_project_post_type');
+		printf('<div class="prdwc-fields" style="margin-bottom:1rem">');
 
 		$fields[] = (empty($project_post_type)) ? array(
 			'name' => 'project-name',
-			'label' => __('Project', 'project-products'),
+			'label' => __('Project', 'project-donations-wc'),
 			'required' => true,
 			'value' => $project,
-			'placeholder' => __("Enter a project name", 'project-products'),
+			'placeholder' => __("Enter a project name", 'project-donations-wc'),
 		) : array(
 			'name' => 'project-id',
-			'label' => __('Project', 'project-products'),
+			'label' => __('Project', 'project-donations-wc'),
 			'required' => true,
 			'type' => 'select',
 			'value' => $project_id,
-			'options' => PROPRO::select_post_options(array(
+			'options' => PRDWC::select_post_options(array(
 				'post_type' => $project_post_type,
 				'orderby'     => 'post_title',
 			)),
 		);
 
-		if(propro_allow_custom_amount()) {
+		if(prdwc_allow_custom_amount()) {
 			$fields[] = array(
 				'name' => 'amount',
 				'label' => ( ($price > 0 ) ? __('Add to fee', 'lodgify-link') : __('Amount', 'lodgify-link') ) . ' (' . get_woocommerce_currency_symbol() . ')',
@@ -224,36 +224,36 @@ class PROPRO {
 				'required' => true,
 				'value' => $amount,
 				'custom_attributes' => array(
-					'min' => get_option('propro_minimum_amount', 1),
+					'min' => get_option('prdwc_minimum_amount', 1),
 					'step' => 'any',
 				)
 			);
 		}
 
 		foreach($fields as $field) {
-			PROPRO::display_custom_field($field);
+			PRDWC::display_custom_field($field);
 		}
 
 
 		// printf(
-    //   '<div class="propro-field propro-field-project-name">
+    //   '<div class="prdwc-field prdwc-field-project-name">
 		// 		<p class="form-row form-row-wide">
-    //     	<label for="propro-project-name" class="required">%s%s</label>
-    //       <input type="%s" class="input-text" name="propro-project-name" value="%s" placeholder="%s" required>
+    //     	<label for="prdwc-project-name" class="required">%s%s</label>
+    //       <input type="%s" class="input-text" name="prdwc-project-name" value="%s" placeholder="%s" required>
     //     </p>
     //   </div>',
-    //   __('Project', 'project-products'),
+    //   __('Project', 'project-donations-wc'),
     //   (empty($project)) ? ' <abbr class="required" title="required">*</abbr>' : ': <span class=project_name>' . $project . '</span>',
     //   (empty($project)) ? 'text' : 'hidden',
     //   $project,
-    //   __("Enter a project name", 'project-products'),
+    //   __("Enter a project name", 'project-donations-wc'),
     // );
 
 		// printf(
-		//   '<div class="propro-field propro-field-amount">
+		//   '<div class="prdwc-field prdwc-field-amount">
 		//     <p class="form-row form-row-wide">
-		//       <label for="propro-amount" class="required">%s%s</label>
-		//       <input type="number" class="input-text" name="propro-amount" value="%s" placeholder="%s" step="any" required>
+		//       <label for="prdwc-amount" class="required">%s%s</label>
+		//       <input type="number" class="input-text" name="prdwc-amount" value="%s" placeholder="%s" step="any" required>
 		//     </p>
 		//   </div>',
 		//   ($price > 0 ) ? __('Add to fee', 'lodgify-link') : __('Amount', 'lodgify-link'),
@@ -265,9 +265,9 @@ class PROPRO {
   }
 
   static function validate_custom_field( $passed, $product_id, $quantity ) {
-    if($passed && propro_is_project_product( $product_id )) {
+    if($passed && prdwc_is_donation( $product_id )) {
 			if(!empty($_REQUEST['project-id'])) $project = $_REQUEST['project-id'];
-      else if(!empty($_POST['propro-project-name'])) $project = $_POST['propro-project-name'];
+      else if(!empty($_POST['prdwc-project-name'])) $project = $_POST['prdwc-project-name'];
       else if(!empty($_REQUEST['project'])) $project = $_REQUEST['project'];
       else $project = NULL;
 			$project = sanitize_text_field($project);
@@ -275,8 +275,8 @@ class PROPRO {
 			$product = wc_get_product( $product_id );
 			$price = $product->get_price();
 
-			if(propro_allow_custom_amount()) {
-				if(!empty($_POST['propro-amount'])) $amount = $_POST['propro-amount'];
+			if(prdwc_allow_custom_amount()) {
+				if(!empty($_POST['prdwc-amount'])) $amount = $_POST['prdwc-amount'];
 				else if(!empty($_REQUEST['amount'])) $amount = $_REQUEST['amount'];
 				else $amount = 0;
 				$amount = sanitize_text_field($amount);
@@ -294,7 +294,7 @@ class PROPRO {
         $product_title = wc_get_product( $product_id )->get_title();
 
         wc_add_notice( sprintf(
-          __('"%s" could not be added to the cart. Please provide a project name.', 'project-products'),
+          __('"%s" could not be added to the cart. Please provide a project name.', 'project-donations-wc'),
           sprintf('<a href="%s">%s</a>', get_permalink($product_id), $product_title),
         ), 'error' );
         return false;
@@ -312,23 +312,23 @@ class PROPRO {
   * @param Boolean $quantity Quantity
   */
   static function add_custom_field_item_data( $cart_item_data, $product_id, $variation_id, $quantity ) {
-		if(!propro_is_project_product( wc_get_product( $product_id ) )) return $cart_item_data;
+		if(!prdwc_is_donation( wc_get_product( $product_id ) )) return $cart_item_data;
 
 		if(!empty($_REQUEST['project-id'])) {
 			$project = get_the_title($_REQUEST['project-id']);
-			if(!empty($project)) $cart_item_data['propro-project-id'] = sanitize_text_field($_REQUEST['project-id']);
+			if(!empty($project)) $cart_item_data['prdwc-project-id'] = sanitize_text_field($_REQUEST['project-id']);
 		}
-    else if(!empty($_POST['propro-project-name'])) $project = $_POST['propro-project-name'];
+    else if(!empty($_POST['prdwc-project-name'])) $project = $_POST['prdwc-project-name'];
     else if(!empty($_REQUEST['project'])) $project = $_REQUEST['project'];
 		else $project = NULL;
-		$cart_item_data['propro-project-name'] = sanitize_text_field($project);
+		$cart_item_data['prdwc-project-name'] = sanitize_text_field($project);
 
-		if(propro_allow_custom_amount()) {
-			if(!empty($_POST['propro-amount'])) $amount = $_POST['propro-amount'];
+		if(prdwc_allow_custom_amount()) {
+			if(!empty($_POST['prdwc-amount'])) $amount = $_POST['prdwc-amount'];
 			else if(!empty($_REQUEST['amount'])) $amount = $_REQUEST['amount'];
 			else if(!empty($_REQUEST['nyp'])) $amount = $_REQUEST['nyp'];
 			else $amount = NULL;
-			$cart_item_data['propro-amount'] = sanitize_text_field($amount);
+			$cart_item_data['prdwc-amount'] = sanitize_text_field($amount);
 		}
 
     return $cart_item_data;
@@ -340,11 +340,11 @@ class PROPRO {
   */
   static function cart_item_name( $name, $cart_item, $cart_item_key ) {
 
-    if( isset( $cart_item['propro-project-name'] ) ) {
+    if( isset( $cart_item['prdwc-project-name'] ) ) {
       $name = sprintf(
-      '%s <span class=propro-project-name>"%s"</span>',
+      '%s <span class=prdwc-project-name>"%s"</span>',
       $name,
-      esc_html( $cart_item['propro-project-name'] ),
+      esc_html( $cart_item['prdwc-project-name'] ),
       );
     }
     return $name;
@@ -355,17 +355,17 @@ class PROPRO {
   */
   static function add_custom_data_to_order( $item, $cart_item_key, $values, $order ) {
     foreach( $item as $cart_item_key=>$values ) {
-      if( isset( $values['propro-project-name'] ) ) {
-        $item->add_meta_data( __( 'Project', 'project-products' ), $values['propro-project-name'], true );
-				if( isset( $values['propro-project-id'] ) )
-				$item->add_meta_data( __( 'Project ID', 'project-products' ), $values['propro-project-id'], true );
+      if( isset( $values['prdwc-project-name'] ) ) {
+        $item->add_meta_data( __( 'Project', 'project-donations-wc' ), $values['prdwc-project-name'], true );
+				if( isset( $values['prdwc-project-id'] ) )
+				$item->add_meta_data( __( 'Project ID', 'project-donations-wc' ), $values['prdwc-project-id'], true );
       }
     }
   }
 
   static function get_price_html( $price_html, $product ) {
     if($product->get_meta( '_linkproject' ) == 'yes') {
-      $price = max($product->get_price(), get_option('propro-project-minimum_price', 0));
+      $price = max($product->get_price(), get_option('prdwc-project-minimum_price', 0));
       if( $price == 0 ) {
         $price_html = apply_filters( 'woocommerce_empty_price_html', '', $product );
       } else {
@@ -375,18 +375,18 @@ class PROPRO {
         } else {
           $price = wc_price( $price ) . $product->get_price_suffix();
         }
-        $price_html = sprintf('<span class="from">%s </span>', __('From', 'project-products')) . $price;
+        $price_html = sprintf('<span class="from">%s </span>', __('From', 'project-donations-wc')) . $price;
       }
     }
     return $price_html;
   }
 
 	static function get_variation_price_html( $data, $product, $variation ) {
-		if(!propro_allow_custom_amount()) return $data;
-		// if(propro_is_project_product( $product->get_id() ))
+		if(!prdwc_allow_custom_amount()) return $data;
+		// if(prdwc_is_donation( $product->get_id() ))
 		if($data['display_price'] > 0) {
 			$data['price_html'] = sprintf(
-				__('%s will be added to the chosen amount, the total price will be calculated before checkout.', 'project-products'),
+				__('%s will be added to the chosen amount, the total price will be calculated before checkout.', 'project-donations-wc'),
 				wp_strip_all_tags($data['price_html']),
 			);
 		} else {
@@ -405,20 +405,20 @@ class PROPRO {
     if ( is_admin() && ! defined( 'DOING_AJAX' ) ) {
       return;
     }
-		if(!propro_allow_custom_amount()) return;
+		if(!prdwc_allow_custom_amount()) return;
 
     // Iterate through each cart item
     foreach( $cart->get_cart() as $cart_key => $cart_item ) {
-      $cached = wp_cache_get('propro_cart_item_processed_' . $cart_key, 'project-products');
+      $cached = wp_cache_get('prdwc_cart_item_processed_' . $cart_key, 'project-donations');
       if(!$cached) {
-        if( isset($cart_item['propro-amount']) &! isset($cart_item['propro-amount_added']) ) {
-					// $cart_item['data']->adjust_price( $cart_item['propro-amount'] );
+        if( isset($cart_item['prdwc-amount']) &! isset($cart_item['prdwc-amount_added']) ) {
+					// $cart_item['data']->adjust_price( $cart_item['prdwc-amount'] );
           $price = (float)$cart_item['data']->get_price( 'edit' );
-          $total = isset($cart_item['propro-amount']) ? $price + $cart_item['propro-amount'] : $price;
+          $total = isset($cart_item['prdwc-amount']) ? $price + $cart_item['prdwc-amount'] : $price;
           $cart_item['data']->set_price( ( $total ) );
-          $cart_item['propro-amount_added'] = true;
+          $cart_item['prdwc-amount_added'] = true;
         }
-        wp_cache_set('propro_cart_item_processed_' . $cart_key, true, 'project-products');
+        wp_cache_set('prdwc_cart_item_processed_' . $cart_key, true, 'project-donations');
       }
     }
   }
@@ -443,4 +443,4 @@ class PROPRO {
 
 }
 
-PROPRO::init();
+PRDWC::init();