+++ includes/post-type-project.php
@@ -1,18 +1,18 @@
-<?php defined( 'PROPRO_VERSION' ) || die;
+<?php defined( 'PRDWC_VERSION' ) || die;
 
-function propro_register_project_posttype() {
+function prdwc_register_project_posttype() {
 	$labels = array(
-		'name'               => esc_html__( 'Projects', 'project-products' ),
-		'singular_name'      => esc_html__( 'Project', 'project-products' ),
-		'add_new'            => esc_html__( 'Add New', 'project-products' ),
-		'add_new_item'       => esc_html__( 'Add New Project', 'project-products' ),
-		'edit_item'          => esc_html__( 'Edit Project', 'project-products' ),
-		'new_item'           => esc_html__( 'New Project', 'project-products' ),
-		'all_items'          => esc_html__( 'All Projects', 'project-products' ),
-		'view_item'          => esc_html__( 'View Project', 'project-products' ),
-		'search_items'       => esc_html__( 'Search Projects', 'project-products' ),
-		'not_found'          => esc_html__( 'Nothing found', 'project-products' ),
-		'not_found_in_trash' => esc_html__( 'Nothing found in Trash', 'project-products' ),
+		'name'               => esc_html__( 'Projects', 'project-donations-wc' ),
+		'singular_name'      => esc_html__( 'Project', 'project-donations-wc' ),
+		'add_new'            => esc_html__( 'Add New', 'project-donations-wc' ),
+		'add_new_item'       => esc_html__( 'Add New Project', 'project-donations-wc' ),
+		'edit_item'          => esc_html__( 'Edit Project', 'project-donations-wc' ),
+		'new_item'           => esc_html__( 'New Project', 'project-donations-wc' ),
+		'all_items'          => esc_html__( 'All Projects', 'project-donations-wc' ),
+		'view_item'          => esc_html__( 'View Project', 'project-donations-wc' ),
+		'search_items'       => esc_html__( 'Search Projects', 'project-donations-wc' ),
+		'not_found'          => esc_html__( 'Nothing found', 'project-donations-wc' ),
+		'not_found_in_trash' => esc_html__( 'Nothing found in Trash', 'project-donations-wc' ),
 		'parent_item_colon'  => '',
 	);
 
@@ -43,18 +43,18 @@ function propro_register_project_posttype() {
 	register_post_type( 'project', apply_filters( 'et_project_posttype_args', $args ) );
 
 	$labels = array(
-		'name'              => esc_html__( 'Project Categories', 'project-products' ),
-		'singular_name'     => esc_html__( 'Project Category', 'project-products' ),
-		'search_items'      => esc_html__( 'Search Categories', 'project-products' ),
-		'all_items'         => esc_html__( 'All Categories', 'project-products' ),
-		'parent_item'       => esc_html__( 'Parent Category', 'project-products' ),
-		'parent_item_colon' => esc_html__( 'Parent Category:', 'project-products' ),
-		'edit_item'         => esc_html__( 'Edit Category', 'project-products' ),
-		'update_item'       => esc_html__( 'Update Category', 'project-products' ),
-		'add_new_item'      => esc_html__( 'Add New Category', 'project-products' ),
-		'new_item_name'     => esc_html__( 'New Category Name', 'project-products' ),
-		'menu_name'         => esc_html__( 'Categories', 'project-products' ),
-		'not_found'         => esc_html__( "You currently don't have any project categories.", 'project-products' ),
+		'name'              => esc_html__( 'Project Categories', 'project-donations-wc' ),
+		'singular_name'     => esc_html__( 'Project Category', 'project-donations-wc' ),
+		'search_items'      => esc_html__( 'Search Categories', 'project-donations-wc' ),
+		'all_items'         => esc_html__( 'All Categories', 'project-donations-wc' ),
+		'parent_item'       => esc_html__( 'Parent Category', 'project-donations-wc' ),
+		'parent_item_colon' => esc_html__( 'Parent Category:', 'project-donations-wc' ),
+		'edit_item'         => esc_html__( 'Edit Category', 'project-donations-wc' ),
+		'update_item'       => esc_html__( 'Update Category', 'project-donations-wc' ),
+		'add_new_item'      => esc_html__( 'Add New Category', 'project-donations-wc' ),
+		'new_item_name'     => esc_html__( 'New Category Name', 'project-donations-wc' ),
+		'menu_name'         => esc_html__( 'Categories', 'project-donations-wc' ),
+		'not_found'         => esc_html__( "You currently don't have any project categories.", 'project-donations-wc' ),
 	);
 
 	$project_category_args = array(
@@ -68,17 +68,17 @@ function propro_register_project_posttype() {
 	register_taxonomy( 'project_category', array( 'project' ), $project_category_args );
 
 	$labels = array(
-		'name'              => esc_html__( 'Project Tags', 'project-products' ),
-		'singular_name'     => esc_html__( 'Project Tag', 'project-products' ),
-		'search_items'      => esc_html__( 'Search Tags', 'project-products' ),
-		'all_items'         => esc_html__( 'All Tags', 'project-products' ),
-		'parent_item'       => esc_html__( 'Parent Tag', 'project-products' ),
-		'parent_item_colon' => esc_html__( 'Parent Tag:', 'project-products' ),
-		'edit_item'         => esc_html__( 'Edit Tag', 'project-products' ),
-		'update_item'       => esc_html__( 'Update Tag', 'project-products' ),
-		'add_new_item'      => esc_html__( 'Add New Tag', 'project-products' ),
-		'new_item_name'     => esc_html__( 'New Tag Name', 'project-products' ),
-		'menu_name'         => esc_html__( 'Tags', 'project-products' ),
+		'name'              => esc_html__( 'Project Tags', 'project-donations-wc' ),
+		'singular_name'     => esc_html__( 'Project Tag', 'project-donations-wc' ),
+		'search_items'      => esc_html__( 'Search Tags', 'project-donations-wc' ),
+		'all_items'         => esc_html__( 'All Tags', 'project-donations-wc' ),
+		'parent_item'       => esc_html__( 'Parent Tag', 'project-donations-wc' ),
+		'parent_item_colon' => esc_html__( 'Parent Tag:', 'project-donations-wc' ),
+		'edit_item'         => esc_html__( 'Edit Tag', 'project-donations-wc' ),
+		'update_item'       => esc_html__( 'Update Tag', 'project-donations-wc' ),
+		'add_new_item'      => esc_html__( 'Add New Tag', 'project-donations-wc' ),
+		'new_item_name'     => esc_html__( 'New Tag Name', 'project-donations-wc' ),
+		'menu_name'         => esc_html__( 'Tags', 'project-donations-wc' ),
 	);
 
 	$project_tag_args = array(