+++ lib/package-updater.php
@@ -6,8 +6,8 @@ if ( ! defined( 'WPINC' ) ) {
 
 /** Enable plugin updates **/
 require_once plugin_dir_path( dirname(__FILE__) ) . 'lib/wp-package-updater/class-wp-package-updater.php';
-$propro_updater = new WP_Package_Updater(
+$prdwc_updater = new WP_Package_Updater(
 	'https://magiiic.com',
-	wp_normalize_path( plugin_dir_path( dirname(__FILE__) ) . "/project-products.php" ),
+	wp_normalize_path( plugin_dir_path( dirname(__FILE__) ) . "/project-donations.php" ),
 	wp_normalize_path( plugin_dir_path( dirname(__FILE__) ) )
 );