+++ CHANGELOG.md
@@ -5,8 +5,8 @@
 
 ### 1.3
 * new choice to create "project" post type or choose an existing post type as projects * product page display a project selection menu  post type if project post type is set, free type text input instead
-* added allow custom amount option for project products
-* added WooCommerce settings tab for Project Products
+* added allow custom amount option for project donations
+* added WooCommerce settings tab for Project Donations
 * added WooCommerce settings tab
 
 ### 1.2.1
@@ -27,7 +27,7 @@
 
 ### 1.0
 * new Project name field on products, autofill if URL parameter project is provided, project name added in cart item name
-* added "Project Product" option to product type section in edit page
+* added "Project Donation" option to product type section in edit page
 * added french localization
 
 ### 0.1.0